requirements.txt
# Python dependencies for MODIS Land Cover Analysis System

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# Statistical analysis
scipy>=1.9.0

# Visualization
matplotlib>=3.6.0
cartopy>=0.21.0

# Configuration
PyYAML>=6.0
